Modularize the packaging code
